+++ tests/src/test/scala/at/logic/gapt/provers/leancop/leancopTest.scala
@@ -1,8 +1,8 @@
 package at.logic.gapt.provers.leancop
 
+import at.logic.gapt.examples.BussTautology
 import at.logic.gapt.expr._
-import at.logic.gapt.proofs.{ Sequent, HOLSequent }
-
+import at.logic.gapt.proofs.{ HOLSequent, Sequent }
 import org.specs2.mutable._
 
 class LeanCoPProverTest extends Specification {
@@ -39,8 +39,13 @@ class LeanCoPProverTest extends Specification {
       LeanCoP.getExpansionProof( seq ) must beSome
     }
 
-    //    "validate the buss tautology for n=1" in { leanCoP.isValid( BussTautology( 1 ) ) must beTrue }
+    "validate the buss tautology for n=2" in { LeanCoP.isValid( BussTautology( 2 ) ) must beTrue }
 
-    // top/bottom cannot be parsed yet
+    "not prove a or b" in { LeanCoP getExpansionProof hof"a | b" must beNone }
+
+    "prove top" in { LeanCoP.getLKProof( Sequent() :+ Top() ) must beSome }
+    "not prove bottom" in { LeanCoP.getLKProof( Sequent() :+ Bottom() ) must beNone }
+    "not refute top" in { LeanCoP.getLKProof( Top() +: Sequent() ) must beNone }
+    "refute bottom" in { LeanCoP.getLKProof( Bottom() +: Sequent() ) must beSome }
   }
 }